requirements.txt
cycler==0.10.0
dj-database-url==0.4.2
Django==1.11.3
django-rest-framework-mongoengine==3.3.1
django-webpack-loader==0.5.0
djangorestframework==3.6.3
functools32==3.2.3.post2
gunicorn==19.7.1
mongoengine==0.13.0
Pygments==2.2.0
pymongo==3.4.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.2
six==1.10.0
subprocess32==3.2.7
whitenoise==3.3.0
numpy== 1.11.1
scipy==0.18.0
matplotlib==1.5.3
blinker==1.4
neutron_detector_eff_functions==1.3.9
